pyproject.toml
[tool.poetry]
name = "docgenai"
version = "0.1.0"
description = "AI-powered documentation and diagramming tool."
authors = ["AI Assistant <assistant@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12.0"
click = "^8.1.7"
transformers = "4.36.2"
torch = "^2.2.0"
jinja2 = "^3.1.2"
PyYAML = "^6.0.1"
python-dotenv = "^1.0.0"
pillow = "^11.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
docgenai = "docgenai.cli:cli"
